broke out the TwitterBot prototype methods into separate files - one to hold general methods used to handle tweets and bots, the other to contain the methods for the streaming api routing
